backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0

# Spark dependencies
pyspark==3.4.1

# Utilities
python-dotenv==1.0.0